Shared layout for auth route group
