+++ frontend/foodieland-app/app/(auth)/page.module.css
@@ -11,35 +11,6 @@
     justify-content: space-between;
 }
 
-.loginRight {
-    flex: 0 1 50%;
-    padding: 15px;
-}
-.loginRightContainer {
-    padding: 30px;
-    background-color: #E7FAFE;
-    border-radius: 20px;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-}
-.loginRightContent {
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-    max-width: 400px;
-    text-align: right;
-}
-
-.loginRightContent > p {
-    font-size: 26px;
-    font-weight: bold;
-    color: #666666;
-    letter-spacing: 1px;
-    line-height: 100%;
-}
-
 .loginHeader {
     padding: 20px;
 }
