+++ frontend/foodieland-app/components/index.ts
@@ -1,7 +1,7 @@
 import Logo from "@/components/Logo";
 import Button from "@/components/Button";
-import LoginForm from "@/components/LoginForm";
-import RegisterForm from "@/components/RegisterForm";
+import LoginForm from "@/components/forms/LoginForm";
+import RegisterForm from "@/components/forms/RegisterForm";
 import Promosection from "@/components/Promosection";
 
 export {
