+++ frontend/foodieland-app/components/promo.module.css
@@ -1,3 +1,17 @@
+.loginRight {
+    flex: 0 1 50%;
+    padding: 15px;
+}
+.loginRightContainer {
+    padding: 30px;
+    background-color: #E7FAFE;
+    border-radius: 20px;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+}
+
 .container {
     padding: 20px;
     display: flex;
